tools/url_classifier/requirements.txt
openai>=1.30.0
pandas>=2.0.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
requests>=2.31.0
